pyproject.toml
[tool.poetry]
name = "esw-utility-scripts"
version = "0.1.0"
description = ""
license = "AGPL-3.0-or-later"
authors = ["m3adow <1872304+m3adow@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "esw_utility_scripts"}]

[tool.poetry.scripts]
image_resizer = "esw_utility_scripts.image_resizer:main"

[tool.poetry.dependencies]
python = ">3.8"
pytomlpp = "^1.0.11"
pillow = "^9.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
